refactor: Define a base Question settings class
