+++ src/argilla_sdk/settings/_common.py
@@ -55,35 +55,3 @@ class SettingsPropertyBase(Resource):
         serialized_model = super().serialize()
         serialized_model["type"] = self.type
         return serialized_model
-
-    ##############################
-    #  Private methods
-    ##############################
-
-    @staticmethod
-    def _render_values_as_options(values: Union[List[str], List[int], Dict[str, str]]) -> List[Dict[str, str]]:
-        """Render values as options for the question so that the model conforms to the API schema"""
-        if isinstance(values, dict):
-            return [{"text": value, "value": key} for key, value in values.items()]
-        elif isinstance(values, list) and all(isinstance(value, str) for value in values):
-            return [{"text": label, "value": label} for label in values]
-        elif isinstance(values, list) and all(isinstance(value, int) for value in values):
-            return [{"value": value} for value in values]
-        else:
-            raise ValueError("Invalid labels format. Please provide a list of strings or a list of dictionaries.")
-
-    @staticmethod
-    def _render_options_as_values(options: List[dict]) -> Dict[str, str]:
-        """Render options as values for the question so that the model conforms to the API schema"""
-        values = {}
-        for option in options:
-            if "text" in option:
-                values[option["value"]] = option["text"]
-            else:
-                values[option["value"]] = option["value"]
-        return values
-
-    @staticmethod
-    def _render_options_as_labels(options: List[Dict[str, str]]) -> List[str]:
-        """Render values as labels for the question so that they can be returned as a list of strings"""
-        return list(SettingsPropertyBase._render_options_as_values(options=options).keys())
